+++ apis/python/node/Cargo.toml
@@ -13,6 +13,7 @@ pyo3 = { version = "0.16", features = ["eyre", "abi3-py37"] }
 eyre = "0.6"
 serde_yaml = "0.8.23"
 flume = "0.10.14"
+dora-runtime = { path = "../../../binaries/runtime" }
 
 [lib]
 name = "dora"
