+++ binaries/runtime/Cargo.toml
@@ -32,7 +32,7 @@ tokio-stream = "0.1.8"
 zenoh = { git = "https://github.com/eclipse-zenoh/zenoh.git", rev = "79a136e4fd90b11ff5d775ced981af53c4f1071b" }
 zenoh-config = { git = "https://github.com/eclipse-zenoh/zenoh.git", rev = "79a136e4fd90b11ff5d775ced981af53c4f1071b" }
 fern = "0.6.1"
-pyo3 = { version = "0.16.5", features = ["auto-initialize", "eyre", "abi3"] }
+pyo3 = { version = "0.16", features = ["eyre", "abi3-py37"] }
 # pyo3-abi3 flag allow simpler linking. See: https://pyo3.rs/v0.13.2/building_and_distribution.html
 flume = "0.10.14"
 dora-message = { path = "../../libraries/message" }
