+++ examples/python-dataflow/dataflow_without_webcam.yml
@@ -1,6 +1,6 @@
 communication:
-  zenoh:
-    prefix: /example-python-no-webcam-dataflow
+  iceoryx:
+    app_name_prefix: example-python-no-webcam-dataflow
 
 nodes:
   - id: no_webcam
