+++ examples/python-dataflow/plot.py
@@ -24,6 +24,7 @@ class Operator:
 
     def __init__(self):
         self.image = []
+        self.bbox = []
 
     def on_input(
         self,
@@ -45,39 +46,39 @@ class Operator:
 
         elif dora_input["id"] == "bbox" and len(self.image) != 0:
             bboxs = np.frombuffer(dora_input["data"], dtype="float32")
-            bboxs = np.reshape(bboxs, (-1, 6))
-            for bbox in bboxs:
-                [
-                    min_x,
-                    min_y,
-                    max_x,
-                    max_y,
-                    confidence,
-                    label,
-                ] = bbox
-                cv2.rectangle(
-                    self.image,
-                    (int(min_x), int(min_y)),
-                    (int(max_x), int(max_y)),
-                    (0, 255, 0),
-                    2,
-                )
+            self.bboxs = np.reshape(bboxs, (-1, 6))
+        for bbox in self.bboxs:
+            [
+                min_x,
+                min_y,
+                max_x,
+                max_y,
+                confidence,
+                label,
+            ] = bbox
+            cv2.rectangle(
+                self.image,
+                (int(min_x), int(min_y)),
+                (int(max_x), int(max_y)),
+                (0, 255, 0),
+                2,
+            )
 
-                cv2.putText(
-                    self.image,
-                    LABELS[int(label)] + f", {confidence:0.2f}",
-                    (int(max_x), int(max_y)),
-                    font,
-                    0.75,
-                    (0, 255, 0),
-                    2,
-                    1,
-                )
+            cv2.putText(
+                self.image,
+                LABELS[int(label)] + f", {confidence:0.2f}",
+                (int(max_x), int(max_y)),
+                font,
+                0.75,
+                (0, 255, 0),
+                2,
+                1,
+            )
 
-            if CI != "true":
-                cv2.imshow("frame", self.image)
-                if cv2.waitKey(1) & 0xFF == ord("q"):
-                    return DoraStatus.STOP
+        if CI != "true":
+            cv2.imshow("frame", self.image)
+            if cv2.waitKey(1) & 0xFF == ord("q"):
+                return DoraStatus.STOP
 
         return DoraStatus.CONTINUE
